requirements.txt
streamlit==1.34.0
gdown==5.2.0
google-generativeai==0.5.2
streamlit-mic-recorder==0.0.8